structure : folder renamed and refactoring
