+++ nodemon.json
@@ -7,5 +7,5 @@
     "ignore": [
         "src/**/*.spec.ts"  
     ],
-    "exec": "ts-node  --transpile-only ./server.ts"
+    "exec": "ts-node  --transpile-only ./src/server.ts"
 }
